+++ rasa_nlu/training_data/loading.py
@@ -18,6 +18,7 @@ from rasa_nlu.training_data.formats import markdown
 from rasa_nlu.training_data.formats.dialogflow import (
     DIALOGFLOW_AGENT, DIALOGFLOW_PACKAGE, DIALOGFLOW_INTENT,
     DIALOGFLOW_ENTITIES, DIALOGFLOW_ENTITY_ENTRIES, DIALOGFLOW_INTENT_EXAMPLES)
+from rasa_nlu.utils import EndpointConfig
 
 logger = logging.getLogger(__name__)
 
@@ -54,25 +55,31 @@ def load_data(resource_name, language='en'):
     data_sets = [_load(f, language) for f in files]
     data_sets = [ds for ds in data_sets if ds]
     if len(data_sets) == 0:
-        return TrainingData()
+        training_data = TrainingData()
     elif len(data_sets) == 1:
-        return data_sets[0]
+        training_data = data_sets[0]
     else:
-        return data_sets[0].merge(*data_sets[1:])
+        training_data = data_sets[0].merge(*data_sets[1:])
+
+    training_data.validate()
+    return training_data
 
 
-def load_data_from_url(url, language='en'):
-    # type: (Text, Optional[Text]) -> TrainingData
+def load_data_from_endpoint(data_endpoint, language='en'):
+    # type: (EndpointConfig, Optional[Text]) -> TrainingData
     """Load training data from a URL."""
 
-    if not utils.is_url(url):
-        raise requests.exceptions.InvalidURL(url)
+    if not utils.is_url(data_endpoint.url):
+        raise requests.exceptions.InvalidURL(data_endpoint.url)
     try:
-        response = requests.get(url)
+        response = data_endpoint.request("get")
         response.raise_for_status()
         temp_data_file = utils.create_temporary_file(response.content,
                                                      mode="w+b")
-        return _load(temp_data_file, language)
+        training_data = _load(temp_data_file, language)
+        training_data.validate()
+
+        return training_data
     except Exception as e:
         logger.warning("Could not retrieve training data "
                        "from URL:\n{}".format(e))