+++ rasa_nlu/training_data/training_data.py
@@ -39,7 +39,8 @@ class TrainingData(object):
     def __init__(self,
                  training_examples=None,
                  entity_synonyms=None,
-                 regex_features=None):
+                 regex_features=None,
+                 lookup_tables=None):
         # type: (Optional[List[Message]], Optional[Dict[Text, Text]]) -> None
 
         if training_examples:
@@ -49,8 +50,8 @@ class TrainingData(object):
         self.entity_synonyms = entity_synonyms if entity_synonyms else {}
         self.regex_features = regex_features if regex_features else []
         self.sort_regex_features()
+        self.lookup_tables = lookup_tables if lookup_tables else []
 
-        self.validate()
         self.print_stats()
 
     def merge(self, *others):
@@ -59,10 +60,12 @@ class TrainingData(object):
         training_examples = deepcopy(self.training_examples)
         entity_synonyms = self.entity_synonyms.copy()
         regex_features = deepcopy(self.regex_features)
+        lookup_tables = deepcopy(self.lookup_tables)
 
         for o in others:
             training_examples.extend(deepcopy(o.training_examples))
             regex_features.extend(deepcopy(o.regex_features))
+            lookup_tables.extend(deepcopy(o.lookup_tables))
 
             for text, syn in o.entity_synonyms.items():
                 check_duplicate_synonym(entity_synonyms, text, syn,
@@ -70,7 +73,8 @@ class TrainingData(object):
 
             entity_synonyms.update(o.entity_synonyms)
 
-        return TrainingData(training_examples, entity_synonyms, regex_features)
+        return TrainingData(training_examples, entity_synonyms,
+                            regex_features, lookup_tables)
 
     @staticmethod
     def sanitize_examples(examples):
@@ -207,11 +211,13 @@ class TrainingData(object):
         data_train = TrainingData(
             train,
             entity_synonyms=self.entity_synonyms,
-            regex_features=self.regex_features)
+            regex_features=self.regex_features,
+            lookup_tables=self.lookup_tables)
         data_test = TrainingData(
             test,
             entity_synonyms=self.entity_synonyms,
-            regex_features=self.regex_features)
+            regex_features=self.regex_features,
+            lookup_tables=self.lookup_tables)
         return data_train, data_test
 
     def print_stats(self):