+++ CHANGELOG.rst
@@ -14,6 +14,37 @@ Added
 
 Changed
 -------
+- Training data is now validated after loading from files in ``loading.py`` instead of on initialisation of
+  ``TrainingData`` object
+
+Removed
+-------
+
+Fixed
+-----
+
+[0.13.4] - 2018-09-19
+^^^^^^^^^^^^^^^^^^^^^
+
+Fixed
+-----
+- pinned matplotlib to 2.x (not ready for 3.0 yet)
+- pytest-services since it wasn't used and caused issues on Windows
+
+[0.13.3] - 2018-08-28
+^^^^^^^^^^^^^^^^^^^^^
+
+Added
+-----
+- ``EndpointConfig`` class that handles authenticated requests (ported from Rasa Core)
+- ``DataRouter()`` class supports a ``model_server`` ``EndpointConfig``, which it regularly queries to fetch NLU models
+- this can be used with ``rasa_nlu.server`` with the ``--endpoint`` option (the key for this the model server config is ``model``)
+- docs on model fetching from a URL
+- ability to specify lookup tables in training data
+
+Changed
+-------
+- loading training data from a URL requires an instance of ``EndpointConfig``
 
 - Changed evaluate behaviour to plot two histogram bars per bin.
   Plotting confidence of right predictions in a wine-ish colour
@@ -24,6 +55,7 @@ Removed
 
 Fixed
 -----
+- re-added support for entity names with special characters in markdown format
 
 [0.13.2] - 2018-08-28
 ^^^^^^^^^^^^^^^^^^^^^